Fix performance issue with assemblies query
The visible_for method for Assemblies has performance issues when there are many private users for assemblies. Also, it is not needed anymore, as the HasPrivateUsers versions uses the modified private_spaces version for assemblies. Finally, it is also wrong, as it's including not published assemblies in the result, while the public_spaces doesn't do that.
This commit redefined the method using the version from HasPrivateUsers. It can be reverted after updating to a version of Decidim that removes the visible_for method from the Assembly model.